requirements.txt
requests==1.1.0
simple-pbkdf2==1.0
pycrypto==2.6.1
mock==1.0.1